somefile may help
pins traced to the end of screen may someone need to hook it to logic
analyzier and to kindle to check on signals. we were not able to solder
wires on the end points. kindle does not have space to solder wires.
will make PCB just to find the screen signals. logic file is signals
generated by Controller.